extract authentication into context provider
