+++ constants.js
@@ -1,0 +1,6 @@
+module.exports = {
+  API_HOST:
+    process.env.NODE_ENV === "production"
+      ? process.env.PROD_API_HOST
+      : process.env.DEV_API_HOST,
+};
