+++ package.json
@@ -11,13 +11,14 @@
   "dependencies": {
     "autoprefixer": "^10.2.4",
     "axios": "^0.21.1",
+    "cookie": "^0.4.1",
     "formik": "^2.2.6",
-    "yup": "0.27.0",
     "next": "10.0.5",
     "postcss": "^8.2.4",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "tailwindcss": "^2.0.2"
+    "tailwindcss": "^2.0.2",
+    "yup": "0.27.0"
   },
   "devDependencies": {
     "eslint": "^7.19.0",
